Move copy-paste funct to its own doc
